Make ModMode work with TagAPI and VanishNoPacket.
VanishNoPacket is a required plugin, if it is not there
then modmode will not work.

If TagAPI is not present then there will be no green names.

When a player in modmode edits a block, the owner is changed
to modmode_<PlayerName>

The player data loading and saving commands are slightly modified
from the NWS code in WorldNBTStorage.

When a player in ModMode disconnects, it reloads the regular player
data so that it will not overwrite the player data on save.