Version 3.0
This version removes dependence on NMS classes and changes the way player
state is stored.  Player data is now stored in YAML files in
plugins/ModMode/players/.  Player attributes are modified exclusively via
the Bukkit API.

This implementation fixes an IllegalStateException in NMS code when
Player.teleport() is called after a ModMode change.  The effect of that
exception was to break Bukkit.getPlayerExact() and led to chunks not loading
for the player.  The problem was most apparent when transitioning out of ModMode
from The End.

To install, stop the server, set an empty list of moderators in
plugins/ModMode/config.yml and copy world/players/normal_<player>.dat to
<player>.dat for all ModMode-capable staff.