Recover Version 4.0.4 source code.
 * Decompile 4.0.4-SNAPSHOT from production servers.
 * Apply changes to the master branch.
 * Revert `synchronized` and `static` method attributes. Pointless in the
   case of `synchronized`.
 * Drop the use of PaperSpigot-specific API to prevent Phantom spawns.
   Modify the staff member's TIME_SINCE_REST statistic instead.
 * Improve handling of fire ticks and fall damage. These are now serialised
   with other saved state, preventing the abuse of ModMode for e.g.
   cancelling fall damage.
 * Fix handling of missing tracks (avoid NPE).

This version should be not significantly more flawed than the version we
have been running in production for the last year. No effort has yet been
made to fix the problem of Moderators having incorrect permissions. To
fix that requires that the plugin schedule ModMode actions in the event
handler of LuckPerms events fired when LuckPerms completes its track
promotions and demotions.