Version 4.1.1
Fix track handling for LuckPerms 4.x (`luckperms4` branch).
New work for LuckPerms 5.x should go on the `master` branch.

It didn't show up in testing but apparently, as with 5.x,
LuckPerms 4.x also loads tracks aynchronously, which would
account for why ModMode 4.0.4-SNAPSHOT (all of 2019) onwards
has been sometimes failing with permissions - it is caching
`null` as the `Track` reference.

So instead of disabling the plugin when the Track is not
immediately available, we start the ball rolling on the async
Track load when the plugin initialises and hope that it will
be fully loaded by the time the Track is needed.